CLI: move more kind-specific logic into drivers
This is a continuation (and probably conclusion) of d06c0a33.
